+++ noodles-fasta/CHANGELOG.md
@@ -10,6 +10,10 @@
 
   * fasta/record/sequence: Implement `FromIterator<u8>`.
 
+  * fasta/record/sequence: Add iterator to complement a sequence
+    (`Sequence::complement`) ([#86]).
+
+[#86]: https://github.com/zaeleus/noodles/issues/86
 [#87]: https://github.com/zaeleus/noodles/issues/87
 
 ## 0.10.0 - 2022-04-14